requirements.txt
functions-framework==3.*
google-cloud-storage==2.11.0
google-cloud-bigquery==3.15.0
pyarrow
pandas